Generate fit diagnostic plots

Generate fit diagnostic plots for each ROI: EICFit the raw data and detected feature fit, rtPeakwidthVert detected peaks retention time apex and peakwidth (vertical and no run order), rtPeakwidthHorzRunOrder detected peaks retention time apex and peakwidth by run order, mzPeakwidthHorzRunOrder detected peaks m/z apex and peakwidth by run order, areaRunOrder detected peaks area by run order, rtHistogram histogram of detected peaks retention time, mzHistogram histogram of detected peaks m/z, areaHistogram histogram of detected peaks area.

# S4 method for peakPantheRAnnotation
annotationDiagnosticPlots(
    object,
    sampleColour,
    sampling,
    verbose
)

Arguments

object

(peakPantheRAnnotation) Annotated peakPantheRAnnotation object

sampleColour

(str) NULL or vector colour for each sample

sampling

(int) Number of points to employ when plotting fittedCurve

verbose

(bool) if TRUE message the plot generation progress

Value

A list (one list per compound) of diagnostic plots: result[[i]]$EICFit, result[[i]]$rtPeakwidthVert, result[[i]]$rtPeakwidthHorzRunOrder, result[[i]]$mzPeakwidthHorzRunOrder, result[[i]]$areaRunOrder, result[[i]]$rtHistogram, result[[i]]$mzHistogram, result[[i]]$areaHistogram, result[[i]]$title

Examples

if(requireNamespace('faahKO')){
## Initialise a peakPantheRAnnotation object with 3 samples and 2 targeted
## compounds

# Paths to spectra files
library(faahKO)
spectraPaths <- c(system.file('cdf/KO/ko15.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko16.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko18.CDF', package = 'faahKO'))

# targetFeatTable
targetFeatTable <- data.frame(matrix(vector(), 2, 8, dimnames=list(c(),
                    c('cpdID','cpdName','rtMin','rt','rtMax','mzMin','mz',
                    'mzMax'))), stringsAsFactors=FALSE)
targetFeatTable[1,] <- c('ID-1', 'Cpd 1', 3310., 3344.888, 3390., 522.194778,
                        522.2, 522.205222)
targetFeatTable[2,] <- c('ID-2', 'Cpd 2', 3280., 3385.577, 3440., 496.195038,
                        496.2, 496.204962)
targetFeatTable[,c(3:8)] <- vapply(targetFeatTable[,c(3:8)], as.numeric,
                                    FUN.VALUE=numeric(2))

emptyAnnotation <- peakPantheRAnnotation(spectraPaths=spectraPaths,
                                        targetFeatTable=targetFeatTable)

annotationDiagnosticPlots(emptyAnnotation)
# Warning: the object has not been annotated, return an empty diagnostic plot
# list
# [[1]]
# NULL
# [[2]]
# NULL
}
#> Warning: the object has not been annotated, return an empty diagnostic plot list
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#>