UI show annotation helper - UI sidebar string

Return a text description of an annotation for UI presentation

annotation_showText_UI_helper(annotProp)

Arguments

annotProp
(list) Named list of annotation properties as created by annotation_showMethod_UI_helper()

Value

(str) Textual description of the annotation to show on UI

Examples

# Input
properties_default <- list(nbCompounds = 0,
                            nbSamples = 0,
                            uROIExist = FALSE,
                            useUROI = FALSE,
                            useFIR = FALSE,
                            isAnnotated = FALSE)

# Generate description
annotation_showText_UI_helper(properties_default)
#> [[1]]
#> [1] "Not annotated"
#> 
#> [[2]]
#> [1] "0 compounds"
#> 
#> [[3]]
#> [1] "0 samples"
#> 
#> [[4]]
#> [1] "updated ROI do not exist (uROI)"
#> 
#> [[5]]
#> [1] "does not use updated ROI (uROI)"
#> 
#> [[6]]
#> [1] "does not use fallback integration regions (FIR)"
#> 
# [[1]]
# [1] "Not annotated"
#
# [[2]]
# [1] "0 compounds"
#
# [[3]]
# [1] "0 samples"
#
# [[4]]
# [1] "updated ROI do not exist (uROI)"
#
# [[5]]
# [1] "does not use updated ROI (uROI)"
#
# [[6]]
# [1] "does not use fallback integration regions (FIR)"