Save to disk all annotation results as csv files

Save to disk all annotation results as annotationName_ ... .csv files: compound metadata (cpdMetadata, cpdID, cpdName) and spectra metadata (spectraMetadata, acquisitionTime, TIC), summary of fit (ratio of peaks found: ratio_peaks_found, ratio of peaks filled: ratio_peaks_filled, mean ppm_error: ppm_error, mean rt_dev_sec: rt_dev_sec), and a file for each column of peakTables (with samples as rows and compounds as columns)

# S4 method for peakPantheRAnnotation
outputAnnotationResult(
    object,
    saveFolder,
    annotationName = "annotationResult",
    verbose = TRUE
)

Arguments

object

(peakPantheRAnnotation) Annotated peakPantheRAnnotation object

saveFolder

(str) Path of folder where the annotation result csv will be saved

annotationName

(str) name of annotation to use in the saved csv

verbose

(bool) If TRUE message progress

Value

None

Examples

if(requireNamespace('faahKO')){
## Initialise a peakPantheRAnnotation object with 3 samples and 2 targeted
## compounds

# Paths to spectra files
library(faahKO)
spectraPaths <- c(system.file('cdf/KO/ko15.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko16.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko18.CDF', package = 'faahKO'))

# targetFeatTable
targetFeatTable <- data.frame(matrix(vector(), 2, 8, dimnames=list(c(),
                    c('cpdID','cpdName','rtMin','rt','rtMax','mzMin','mz',
                    'mzMax'))), stringsAsFactors=FALSE)
targetFeatTable[1,] <- c('ID-1', 'Cpd 1', 3310., 3344.888, 3390., 522.194778,
                        522.2, 522.205222)
targetFeatTable[2,] <- c('ID-2', 'Cpd 2', 3280., 3385.577, 3440., 496.195038,
                        496.2, 496.204962)
targetFeatTable[,c(3:8)] <- vapply(targetFeatTable[,c(3:8)], as.numeric,
                                    FUN.VALUE=numeric(2))

emptyAnnotation <- peakPantheRAnnotation(spectraPaths=spectraPaths,
                                        targetFeatTable=targetFeatTable)

# Calculate annotation
annotation      <- peakPantheR_parallelAnnotation(emptyAnnotation, ncores=0,
                                getAcquTime=FALSE, verbose=FALSE)$annotation

# temporary location
savePath1       <- tempdir()
outputAnnotationResult(annotation, saveFolder=savePath1,
                        annotationName='testProject', verbose=TRUE)
}
#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.
#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.
#> Polarity can not be extracted from netCDF files, please set manually the polarity with the 'polarity' method.
#> Compound metadata saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_cpdMetadata.csv
#> Spectra metadata saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_spectraMetadata.csv
#> Peak measurement "found" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_found.csv
#> Peak measurement "rtMin" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_rtMin.csv
#> Peak measurement "rt" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_rt.csv
#> Peak measurement "rtMax" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_rtMax.csv
#> Peak measurement "mzMin" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_mzMin.csv
#> Peak measurement "mz" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_mz.csv
#> Peak measurement "mzMax" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_mzMax.csv
#> Peak measurement "peakArea" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_peakArea.csv
#> Peak measurement "peakAreaRaw" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_peakAreaRaw.csv
#> Peak measurement "maxIntMeasured" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_maxIntMeasured.csv
#> Peak measurement "maxIntPredicted" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_maxIntPredicted.csv
#> Peak measurement "is_filled" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_is_filled.csv
#> Peak measurement "ppm_error" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_ppm_error.csv
#> Peak measurement "rt_dev_sec" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_rt_dev_sec.csv
#> Peak measurement "tailingFactor" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_tailingFactor.csv
#> Peak measurement "asymmetryFactor" saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_asymmetryFactor.csv
#> Summary saved at C:\Users\AppData\Local\Temp\RtmpK4H7jx/testProject_summary.csv