peakPantheR Graphical User Interface

peakPantheR Graphical User Interface (GUI) implements all the functions for the parallel detection, integration and reporting of pre-defined features in multiple mass spectrometry data files. To exit press ESC in the command line.

peakPantheR_start_GUI(browser = TRUE)

Arguments

browser

If TRUE open the graphical user interface in a web browser instead of a R window. Default is TRUE

Value

None, start GUI. To exit press ESC in the command line.

Examples

print("Start graphical interface, press 'ESC' in the command line to stop")
#> [1] "Start graphical interface, press 'ESC' in the command line to stop"
# peakPantheR_start_GUI()