Plot samples raw data and detected feature for a single ROI

Plot a ROI across multiple samples (x axis is RT, y axis is intensity) with the matching detected peak rt and peakwidth under it. If curveFit is provided, the fitted curve for each compound is added. RT and peakwidth are plotted in the order spectra are passed, with the first spectra on top.

plotEICDetectedPeakwidth(
    ROIDataPointSampleList,
    cpdID,
    cpdName,
    rt,
    rtMin,
    rtMax,
    mzMin,
    mzMax,
    ratio = 0.85,
    sampling = 250,
    curveFitSampleList = NULL,
    sampleColour = NULL,
    verbose = TRUE
)

Arguments

ROIDataPointSampleList
(list) list of data.frame of raw data points for each sample (retention time 'rt', mass 'mz' and intensity 'int' (as column) of each raw data points (as row)).

cpdID
(str) Compound ID

cpdName
(str) Compound Name

rt
(float) vector of detected peak apex retention time (in sec)

rtMin
(float) vector of detected peak minimum retention time (in sec)

rtMax
(float) vector ofdetected peak maximum retention time (in sec)

mzMin
(float) ROI minimum m/z (matching EIC)

mzMax
(float) ROI maximum m/z (matching EIC)

ratio
(float) value between 0 and 1 defining the vertical percentage taken by the EICs subplot

sampling
(int) Number of points to employ when plotting fittedCurve

curveFitSampleList
(list) NULL or a list of peakPantheR_curveFit (or NA) for each sample

sampleColour
(str) NULL or vector colour for each sample (same length as EICs, rt, rtMin, rtMax)

verbose
(bool) if TRUE message when NA scans are removed

Value

Grob (ggplot object)