Reset a peakPantheRAnnotation and alter samples and compounds information

Reset a peakPantheRAnnotation (remove results and set isAnnotated=FALSE). If a different number of samples ( spectraPaths) or compounds (targetFeatTable) are passed, the object will be initialised to the new size. For input values left as NULL, the slots (filepath (from spectraPaths), ROI, cpdID, cpdName (from targetFeatTable), uROI, FIR, cpdMetadata, spectraMetadata, uROIExist, useUROI and useFIR) will be filled with values from previousAnnotation.

# S4 method for peakPantheRAnnotation
resetAnnotation(
    previousAnnotation,
    spectraPaths = NULL,
    targetFeatTable = NULL,
    uROI = NULL,
    FIR = NULL,
    cpdMetadata = NULL,
    spectraMetadata = NULL,
    uROIExist = NULL,
    useUROI = NULL,
    useFIR = NULL,
    verbose = TRUE,
    ...
)

Arguments

previousAnnotation
(peakPantheRAnnotation) object to reset

spectraPaths
NULL or a character vector of spectra file paths, to set samples to process

targetFeatTable
NULL or a data.frame of compounds to target as rows and parameters as columns: cpdID (str), cpdName (str), rtMin (float in seconds), rt (float in seconds, or NA), rtMax (float in seconds), mzMin (float), mz (float or NA), mzMax (float). Set compounds to target.

uROI
NULL or a data.frame of updated Regions Of Interest (uROI) with compounds as row and uROI parameters as columns: rtMin (float in seconds), rt (float in seconds, or NA), rtMax (float in seconds), mzMin (float), mz (float or NA), mzMax (float).

FIR
NULL or a data.frame of Fallback Integration Regions (FIR) with compounds as row and FIR parameters as columns: rtMin (float in seconds), rtMax (float in seconds), mzMin (float), mzMax (float).

cpdMetadata
NULL or a data.frame of compound metadata, with compounds as row and metadata as columns

spectraMetadata
NULL or a data.frame of sample metadata, with samples as row and metadata as columns

uROIExist
NULL or a logical stating if uROI have been set

useUROI
NULL or a logical stating if uROI are to be used

useFIR
NULL or a logical stating if FIR are to be used

verbose
(bool) If TRUE message progress

...
Additional slots and values to set when resetting the object (cpdID, cpdName, ROI, filepath, TIC, acquisitionTime, peakTables, dataPoints, peakFit)

Value

(peakPantheRAnnotation) object reset with previous results removed and slots updated

Examples

if(requireNamespace('faahKO')){
## Initialise a peakPantheRAnnotation object with 3 samples and 2 targeted
## compounds

# Paths to spectra files
library(faahKO)
spectraPaths <- c(system.file('cdf/KO/ko15.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko16.CDF', package = 'faahKO'))

# targetFeatTable
targetFeatTable <- data.frame(matrix(vector(), 2, 8, dimnames=list(c(),
                    c('cpdID','cpdName','rtMin','rt','rtMax','mzMin','mz',
                    'mzMax'))), stringsAsFactors=FALSE)
targetFeatTable[1,] <- c('ID-1', 'Cpd 1', 3310., 3344.888, 3390., 522.194778,
                        522.2, 522.205222)
targetFeatTable[2,] <- c('ID-2', 'Cpd 2', 3280., 3385.577, 3440., 496.195038,
                        496.2, 496.204962)
targetFeatTable[,c(3:8)] <- vapply(targetFeatTable[,c(3:8)], as.numeric,
                                    FUN.VALUE=numeric(2))

smallAnnotation  <- peakPantheRAnnotation(spectraPaths=spectraPaths, 
                                        targetFeatTable=targetFeatTable)
smallAnnotation
# An object of class peakPantheRAnnotation
#  2 compounds in 2 samples.
#  updated ROI do not exist (uROI)
#  does not use updated ROI (uROI)
#  does not use fallback integration regions (FIR)
#  is not annotated

# Reset and change number of spectra
newSpectraPaths  <- c(system.file('cdf/KO/ko15.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko16.CDF', package = 'faahKO'),
                    system.file('cdf/KO/ko18.CDF', package = 'faahKO'))
largerAnnotation <- resetAnnotation(smallAnnotation,
                                    spectraPaths=newSpectraPaths,
                                    verbose=TRUE)
largerAnnotation
# An object of class peakPantheRAnnotation
#  2 compounds in 3 samples.
#  updated ROI do not exist (uROI)
#  does not use updated ROI (uROI)
#  does not use fallback integration regions (FIR)
#  is not annotated
}

#> peakPantheRAnnotation object being reset:

#>   Previous "ROI", "cpdID" and "cpdName" value kept

#>   Previous "uROI" value kept

#>   Previous "FIR" value kept

#>   Previous "cpdMetadata" value kept

#>   New "spectraPaths" value set

#>   Targeted spectra changed, previous "spectraMetadata" cannot be kept and set to default

#>   Previous "uROIExist" value kept

#>   Previous "useUROI" value kept

#>   Previous "useFIR" value kept

#> An object of class peakPantheRAnnotation
#>  2 compounds in 3 samples. 
#>   updated ROI do not exist (uROI)
#>   does not use updated ROI (uROI)
#>   does not use fallback integration regions (FIR)
#>   is not annotated