Save to disk a plot of all ROI EIC and detected feature range

Plot and save a .png of all ROI (x is RT, y is intensity), with the matching detected peak rt and peakwidth under it.

saveSingleFileMultiEIC(ROIsDataPoint, curveFit, foundPeakTable, savePath,
    width = 15, height = 15, verbose = TRUE)

Arguments

ROIsDataPoint
(list) a list of data.frame of raw data points for each ROI (retention time 'rt', mass 'mz' and intensity 'int' (as column) of each raw data points (as row)).

curveFit
(list) a list of peakPantheR_curveFit or NA for each ROI

foundPeakTable
(data.frame) data.frame as generated by findTargetFeatures, with features as rows and peak properties as columns. The following columns are mandatory: cpdID, cpdName, rt, rtmin, rtmax, mzmin, mzmax.

savePath
(str) Full path to save a .png of all ROI EICs, expect 'filepath/filename.png'.

width
(float) Width in cm for a single ROI plot (if more than one plot in total, 2 columns will be used). dpi set to a 100.

height
(float) height in a cm for a single ROI plot. dpi set to 100

verbose
(bool) if TRUE message progress

Value

None